[arp] Enhancement: Refactoring Module
